+++ dline/deadline.go
@@ -1,4 +1,4 @@
-package deadline
+package dline
 
 import "github.com/filecoin-project/go-state-types/abi"
 
@@ -7,7 +7,7 @@ import "github.com/filecoin-project/go-state-types/abi"
 // Windows are non-overlapping ranges [Open, Close), but the challenge epoch for a window occurs before
 // the window opens.
 // The current epoch may not necessarily lie within the deadline or proving period represented here.
-type DeadlineInfo struct {
+type Info struct {
 	// Deadline parameters
 	CurrentEpoch abi.ChainEpoch // Epoch at which this info was calculated.
 	PeriodStart  abi.ChainEpoch // First epoch of the proving period (<= CurrentEpoch).
@@ -26,64 +26,64 @@ type DeadlineInfo struct {
 }
 
 // Whether the proving period has begun.
-func (d *DeadlineInfo) PeriodStarted() bool {
+func (d *Info) PeriodStarted() bool {
 	return d.CurrentEpoch >= d.PeriodStart
 }
 
 // Whether the proving period has elapsed.
-func (d *DeadlineInfo) PeriodElapsed() bool {
+func (d *Info) PeriodElapsed() bool {
 	return d.CurrentEpoch >= d.NextPeriodStart()
 }
 
 // The last epoch in the proving period.
-func (d *DeadlineInfo) PeriodEnd() abi.ChainEpoch {
+func (d *Info) PeriodEnd() abi.ChainEpoch {
 	return d.PeriodStart + d.WPoStProvingPeriod - 1
 }
 
 // The first epoch in the next proving period.
-func (d *DeadlineInfo) NextPeriodStart() abi.ChainEpoch {
+func (d *Info) NextPeriodStart() abi.ChainEpoch {
 	return d.PeriodStart + d.WPoStProvingPeriod
 }
 
 // Whether the current deadline is currently open.
-func (d *DeadlineInfo) IsOpen() bool {
+func (d *Info) IsOpen() bool {
 	return d.CurrentEpoch >= d.Open && d.CurrentEpoch < d.Close
 }
 
 // Whether the current deadline has already closed.
-func (d *DeadlineInfo) HasElapsed() bool {
+func (d *Info) HasElapsed() bool {
 	return d.CurrentEpoch >= d.Close
 }
 
 // The last epoch during which a proof may be submitted.
-func (d *DeadlineInfo) Last() abi.ChainEpoch {
+func (d *Info) Last() abi.ChainEpoch {
 	return d.Close - 1
 }
 
 // Epoch at which the subsequent deadline opens.
-func (d *DeadlineInfo) NextOpen() abi.ChainEpoch {
+func (d *Info) NextOpen() abi.ChainEpoch {
 	return d.Close
 }
 
 // Whether the deadline's fault cutoff has passed.
-func (d *DeadlineInfo) FaultCutoffPassed() bool {
+func (d *Info) FaultCutoffPassed() bool {
 	return d.CurrentEpoch >= d.FaultCutoff
 }
 
 // Returns the next instance of this deadline that has not yet elapsed.
-func (d *DeadlineInfo) NextNotElapsed() *DeadlineInfo {
+func (d *Info) NextNotElapsed() *Info {
 	next := d
 	for next.HasElapsed() {
-		next = NewDeadlineInfo(next.NextPeriodStart(), d.Index, d.CurrentEpoch, d.WPoStPeriodDeadlines, d.WPoStProvingPeriod, d.WPoStChallengeWindow, d.WPoStChallengeLookback, d.FaultDeclarationCutoff)
+		next = NewInfo(next.NextPeriodStart(), d.Index, d.CurrentEpoch, d.WPoStPeriodDeadlines, d.WPoStProvingPeriod, d.WPoStChallengeWindow, d.WPoStChallengeLookback, d.FaultDeclarationCutoff)
 	}
 	return next
 }
 
 // Returns deadline-related calculations for a deadline in some proving period and the current epoch.
-func NewDeadlineInfo(periodStart abi.ChainEpoch, deadlineIdx uint64, currEpoch abi.ChainEpoch, wPoStPeriodDeadlines uint64, wPoStProvingPeriod, wPoStChallengeWindow, wPoStChallengeLookback, faultDeclarationCutoff abi.ChainEpoch) *DeadlineInfo {
+func NewInfo(periodStart abi.ChainEpoch, deadlineIdx uint64, currEpoch abi.ChainEpoch, wPoStPeriodDeadlines uint64, wPoStProvingPeriod, wPoStChallengeWindow, wPoStChallengeLookback, faultDeclarationCutoff abi.ChainEpoch) *Info {
 	if deadlineIdx < wPoStPeriodDeadlines {
 		deadlineOpen := periodStart + (abi.ChainEpoch(deadlineIdx) * wPoStChallengeWindow)
-		return &DeadlineInfo{
+		return &Info{
 			CurrentEpoch: currEpoch,
 			PeriodStart:  periodStart,
 			Index:        deadlineIdx,
@@ -101,7 +101,7 @@ func NewDeadlineInfo(periodStart abi.ChainEpoch, deadlineIdx uint64, currEpoch a
 	} else {
 		// Return deadline info for a no-duration deadline immediately after the last real one.
 		afterLastDeadline := periodStart + wPoStProvingPeriod
-		return &DeadlineInfo{
+		return &Info{
 			CurrentEpoch: currEpoch,
 			PeriodStart:  periodStart,
 			Index:        deadlineIdx,
